# packages/api/openapi.yaml
openapi: '3.1.0'
info:
  title: Questadi API
  version: '1.0.0'

servers:
  - url: http://localhost:3000

paths:

  # ── Auth ──────────────────────────────────────────────────────────────────

  /auth/otp/request:
    post:
      operationId: requestOtp
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpRequestBody'
      responses:
        '200':
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpSentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'

  /auth/otp/verify:
    post:
      operationId: verifyOtp
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpVerifyBody'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ── Quotes ────────────────────────────────────────────────────────────────

  /user/captures:
    get:
      operationId: getCaptures
      tags: [user]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Quotes captured by the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createCapture
      tags: [user]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuoteRequest'
      responses:
        '201':
          description: Capture created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '200':
          description: Identical capture already exists — existing quote returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /user/captures/{id}:
    delete:
      operationId: deleteCapture
      tags: [user]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QuoteId'
      responses:
        '204':
          description: Capture deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /user/contacts:
    get:
      operationId: searchContacts
      tags: [user]
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 8
      responses:
        '200':
          description: People previously captured by the authenticated user, filtered by name
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AutocompleteResult'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Public (token-gated, no JWT required) ─────────────────────────────────

  /tokens/{token}:
    get:
      operationId: resolveLink
      tags: [public]
      parameters:
        - $ref: '#/components/parameters/Token'
      responses:
        '200':
          description: Quote link resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteLinkPayload'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'
    patch:
      operationId: approveQuotePublic
      tags: [public]
      parameters:
        - $ref: '#/components/parameters/Token'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isPublic]
              properties:
                isPublic:
                  type: boolean
                  enum: [true]
      responses:
        '200':
          description: Quote marked public
          content:
            application/json:
              schema:
                type: object
                required: [isPublic]
                properties:
                  isPublic:
                    type: boolean
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'

  /tokens/{token}/removal-requests:
    post:
      operationId: requestRemoval
      tags: [public]
      parameters:
        - $ref: '#/components/parameters/Token'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  maxLength: 500
      responses:
        '201':
          description: Removal request submitted
          content:
            application/json:
              schema:
                type: object
                required: [requested]
                properties:
                  requested:
                    type: boolean
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'

  /authors/{accountId}/quotes:
    get:
      operationId: getAuthorQuotes
      tags: [public]
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Author's public quote profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorProfile'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    QuoteId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

    Token:
      name: token
      in: path
      required: true
      schema:
        type: string
        minLength: 64
        maxLength: 64

  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not permitted to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Gone:
      description: Resource no longer available (expired or deleted)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string

    OtpRequestBody:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          description: E.164 format, e.g. +14155552671

    OtpSentResponse:
      type: object
      required: [sent]
      properties:
        sent:
          type: boolean

    OtpVerifyBody:
      type: object
      required: [phone, code]
      properties:
        phone:
          type: string
        code:
          type: string
          minLength: 6
          maxLength: 6

    AuthResponse:
      type: object
      required: [token, account]
      properties:
        token:
          type: string
        account:
          $ref: '#/components/schemas/Account'

    Account:
      type: object
      required: [id, displayName, avatarInitials, registered, createdAt]
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        fullName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        contactId:
          type: string
          nullable: true
        avatarInitials:
          type: string
        registered:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastActive:
          type: string
          format: date-time
          nullable: true
        createdBy:
          type: string
          format: uuid
          nullable: true

    AttributedToInput:
      oneOf:
        - title: ExistingAccount
          type: object
          required: [existingId]
          properties:
            existingId:
              type: string
              format: uuid
        - title: PhoneContact
          type: object
          required: [phone, displayName]
          properties:
            phone:
              type: string
              description: E.164 format
            displayName:
              type: string
              minLength: 1
              maxLength: 100
        - title: AnonymousContact
          type: object
          required: [displayName]
          properties:
            displayName:
              type: string
              minLength: 1
              maxLength: 100

    CreateQuoteRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500
        attributedTo:
          $ref: '#/components/schemas/AttributedToInput'

    Quote:
      type: object
      required: [id, text, capturedBy, capturedAt, isPublic]
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        capturedBy:
          type: string
          format: uuid
        attributedTo:
          type: string
          format: uuid
          nullable: true
        capturedAt:
          type: string
          format: date-time
        isPublic:
          type: boolean

    LibraryPersonGroup:
      type: object
      required: [person, quotes, lastQuotedAt]
      properties:
        person:
          $ref: '#/components/schemas/Account'
        quotes:
          type: array
          items:
            $ref: '#/components/schemas/Quote'
        lastQuotedAt:
          type: string
          format: date-time

    LibraryResponse:
      type: object
      required: [people, unattributed]
      properties:
        people:
          type: array
          items:
            $ref: '#/components/schemas/LibraryPersonGroup'
        unattributed:
          type: array
          items:
            $ref: '#/components/schemas/Quote'

    AutocompleteResult:
      type: object
      required: [id, displayName, avatarInitials]
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        fullName:
          type: string
          nullable: true
        avatarInitials:
          type: string
        phone:
          type: string
          nullable: true
        lastQuotedAt:
          type: string
          format: date-time
          nullable: true

    QuoteLinkPayload:
      type: object
      required: [quoteId, quoteText, authorAccountId, authorDisplayName, capturerFirstName, captured_at, is_public, removal_requested, token_expires_at]
      properties:
        quoteId:
          type: string
          format: uuid
        quoteText:
          type: string
        authorAccountId:
          type: string
          format: uuid
        authorDisplayName:
          type: string
        capturerFirstName:
          type: string
        captured_at:
          type: string
          format: date-time
        is_public:
          type: boolean
        removal_requested:
          type: boolean
        token_expires_at:
          type: string
          format: date-time

    AuthorProfile:
      type: object
      required: [accountId, displayName, quotes]
      properties:
        accountId:
          type: string
          format: uuid
        displayName:
          type: string
          nullable: true
        quotes:
          type: array
          items:
            type: object
            required: [id, text, captured_at]
            properties:
              id:
                type: string
                format: uuid
              text:
                type: string
              captured_at:
                type: string
                format: date-time